2_Authorization/requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
Requests==2.31.0
